+++ semana9/astromatch/src/App.js
@@ -1,10 +1,7 @@
 import './App.css';
-import Home from './components/Home/Home';
-import Match from './components/Matches/Match';
-import { useState } from 'react'
-import Box from '@material-ui/core/Box'
 import styled from 'styled-components'
 import Footer from './components/Footer'
+import RenderScreen from './screens/RenderScreen';
 
 
 const Background = styled.div`
@@ -18,38 +15,9 @@ const Background = styled.div`
 
 
 function App() {
-  const [currentPage, setCurrentPage] = useState("home")
-
-  const goToHome = () => {
-    setCurrentPage("home")
-  }
-
-  const goToMatches = () => {
-    setCurrentPage("matches")
-  }
-
-  let renderedPage
-  switch (currentPage) {
-    case "home":
-      renderedPage = <Home currentPage={currentPage} goToHome={goToHome} goToMatches={goToMatches} />
-      break;
-    case "matches":
-      renderedPage = <Match currentPage={currentPage} goToHome={goToHome} goToMatches={goToMatches} />
-      break;
-  }
-
   return (
     <Background>
-      <Box
-        height={600}
-        width={350}
-        borderRadius={8}
-        bgcolor="white"
-        boxShadow={10}
-        overflow="hidden"
-      >
-        {renderedPage}
-      </Box>
+      <RenderScreen />
       <Footer />
     </Background>
