+++ semana9/astromatch/src/constants/colors.js
@@ -1,0 +1,13 @@
+
+// cores usadas apenas para referência para minha cabeça doida
+// #66C7F4
+// #FE039D
+// #62BCAF
+// #E8E8E8
+// #191919
+// #F7D084
+// #FDC9CB
+// #FF1053
+// #606060
+// #FFFFFF
+
