+++ semana9/astromatch/src/constants/urls.js
@@ -1,0 +1,2 @@
+export const baseUrl = "https://us-central1-missao-newton.cloudfunctions.net/astroMatch/nicole"
+
